### 2022-04-01 #168 @whedon
[REVIEW]: pendsim: Developing, Simulating, and Visualizing Feedback Controlled Inverted Pendulum Dynamics
**Submitting author:** @rland93 (<a href="http://orcid.org/0000-0001-5394-2737">Mike Sutherland</a>)
**Repository:** <a href="https://github.com/EASEL-UCI/pendsim" target ="_blank">https://github.com/EASEL-UCI/pendsim</a>
**Version:** v1.2

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
