### 2022-04-01 #168 @moorepants
@cooperrc, @repagh, and @rland93 welcome to the review issue. For the reviewers please read through the above information. I was wrong in the other thread on the timeline, it seems that **6 weeks** is the recommended due date. Note that JOSE reviews are a bit different than most academic journal reviews in that they are open here on Github and there can be a conversation among the reviewers, editors, and authors to arrive a final version that meets the standards for publication. You all can start the review when you'd like and I'll jump in to answer questions or help things move along. Thanks again for volunteering your time and expertise!
