### 2022-04-01 #168 @moorepants
Also I believe this submission is the second type we accept "open-source software, created as educational technology or infrastructure" (see https://openjournals.readthedocs.io/en/jose/submitting.html). @rland93, is that correct?
