### 2022-04-23 #168 @moorepants
@cooperrc and @repagh can we get an update on the status of your reviews? I see that @repagh has worked through his checklist. Do we expect completion of the first reviews within the 6 weeks deadline? Thanks -- Jason
