### 2022-04-23 #168 @cooperrc
> @cooperrc and @repagh can we get an update on the status of your reviews? I see that @repagh has worked through his checklist. Do we expect completion of the first reviews within the 6 weeks deadline? Thanks -- Jason

Thanks for checking in @moorepants. I'm getting back to this today. I'll get my initial thoughts down this week (4/25-4/29). Will have my review done within the 6-week timeframe (before May 13)
