### 2022-05-11 #168 @cooperrc
I added an issue (closed) and two PRs to the `pendsim` repo.

My review is as follows:

The package and the paper are excellent. The one spot that seems to be lacking is a clear path to extend the package (maybe another section for developers). In the paper, @rland93 says the user can implement any control policy with a new class. It would be great to see _how_ to create the new class and _how/if_ the new class could be added to the current repository of controller algorithms.

Another minor note, the `## Example Notebooks` could be added to the documentation [Tutorials](https://rland93.github.io/pendsim/tutorial.html) for more in-depth examples.

- [x] https://github.com/rland93/pendsim/issues/11

- [ ] https://github.com/rland93/pendsim/pull/13

- [ ] https://github.com/rland93/pendsim/pull/14
