### 2022-05-11 #168 @rland93
Will work on implementing these changes this week. Details:

https://github.com/rland93/pendsim/issues/15
https://github.com/rland93/pendsim/issues/16
https://github.com/rland93/pendsim/issues/17

Thanks!
