### 2022-07-21 #168 @moorepants
David Copp is currently listed as a co-author on the paper. Although, I do not see any contributions from him in the pendsim repository: https://github.com/rland93/pendsim/graphs/contributors. We would need his github handle to swap him out, of course.
