### 2022-08-26 #168 @labarba
I don't remember dealing with a change in target repo. @openjournals/dev – can you confirm whether we can edit the submission repository?
