### 2022-08-29 c89bc9b
Update paper.md
added controller design assignment notebook and short discussion on how instructors can use pendsim
