### 2022-08-29 b4244d5
Update paper.md
fixed control assignment notebook link
