### 2022-08-29 dc7aac5
Update paper.md
added notebook link to new example usage discussion
