### 2022-08-29 4aae09a
Update paper.md
added link to docs on how to create a custom controller class
