+++ paper.md
@@ -186,9 +186,8 @@ Several controller implementations are pre-built. These include:
 
 -   Energy Swing-Up Controller
 
-The user can implement any control policy by creating a new class
-[??link to
-docs??](https://github.com/openjournals/jose-reviews/issues/168). Users
+The user can implement any control policy by [creating a new custom 
+controller class]((https://rland93.github.io/pendsim/customctl.html). Users
 can test new open-ended controller designs.  Controllers can dump data
 into the simulation results so that intermediate control inputs are
 accessible to the final results of the simulation.
