+++ README.md
@@ -1,3 +1,12 @@
 # pendsim: an Inverted Pendulum-On-A-Cart Sandbox!
+Created by Mike Sutherland at UC Irvine.
+For documentation, see: http://rland93.github.io/pendsim/
 
-For details, see: http://rland93.github.io/pendsim/
+pendsim is a Python package that can be used to help 
+students design, analyze, and visualize virtual 
+experiments for an inverted pendulum on a cart with 
+feedback control. Example notebooks cover topics 
+ranging from linearization to state estimation and 
+optimal control for this classical nonlinear system.
+
+## Contributing to This Software
