### 2022-09-16 8b2eaa0
Update README.md
added contributing and reporting issues section
