+++ README.md
@@ -1,12 +1,18 @@
 # pendsim: an Inverted Pendulum-On-A-Cart Sandbox!
-Created by Mike Sutherland at UC Irvine.
+Created by Mike Sutherland with the EASEL group at UC Irvine.
 For documentation, see: http://rland93.github.io/pendsim/
 
-pendsim is a Python package that can be used to help 
+*pendsim* is a Python package that can be used to help 
 students design, analyze, and visualize virtual 
 experiments for an inverted pendulum on a cart with 
 feedback control. Example notebooks cover topics 
 ranging from linearization to state estimation and 
 optimal control for this classical nonlinear system.
 
-## Contributing to This Software
+## Contributing and reporting issues
+
+To report any issues using this software, please open a new issue.
+To contribute to this software, please work in a new branch and then create a pull request for the proposed updates. Pull requests will be evaluated and merged as appropriate.
+
+We look forward to incorporating new algorithm examples and notebooks!
+
