### 2022-09-16 fb78f87
Update paper.md
added references section
