### 2022-09-16 #168 @EASEL-UCI
The following changes were made:
- Added a section on "Contributing and reporting issues" to README.md
- Added References section to paper.md
- Added discussion of how the package and notebooks can be used in teaching
- Added an example course assignment notebook on designing a PID controller

@cooperrc: Thank you for your feedback and the excellent proposed changes. I believe we have addressed your comments, but if not or if you have other comments or concerns, please let me know.

@repagh: Thank you for your helpful feedback and for catching errors. Please find our responses to your comments below.

This is the only Python package I am aware of that allows for design, analysis, and visualization of control and estimation algorithms for an inverted pendulum system. Moreover, the example notebooks cover multiple algorithms, whereas other projects may focus solely on PID control, for example. We also added a tutorial on how to create a custom control algorithm (https://rland93.github.io/pendsim/customctl.html).

Regarding how to use this in teaching, you make an excellent point. We added the following paragraph in the Example Usage section to briefly describe how this package can be used in education, and we added an example assignment notebook. The emphasis here is on augmenting in-class activities or in-lab hardware experiments rather than serving as the sole instructional material on a topic. For example, there are many excellent online resources for learning PID control; we do not try to replicate those. Instead we aim to provide a sandbox that students can easily use to gain experience tuning a PID controller, visualizing the results, and reflecting on the process.

"This design-test-visualize sequence also allows instructors to introduce
students to new topics and create interactive assignments that can augment
theoretical class discussions and hardware experiments. Thus, `pendsim`
may be naturally used in engineering courses related to dynamics and control
that are taught in any format, including in-person, hybrid, and online settings.
An example [PID Controller Design Assignment notebook](https://colab.research.google.com/github/EASEL-UCI/pendsim/blob/main/notebooks/PID_controller_design_assignment.ipynb) is
available to show how students may use `pendsim` in a course assignment."
