### 2022-10-10 #168 @moorepants
@EASEL-UCI I see some math rendering errors here:
![image](https://user-images.githubusercontent.com/276007/194847314-63ffcdf8-7397-4b05-bcdf-e2c42bb45af2.png)
https://rland93.github.io/pendsim/customctl.html#Model-Predictive-Control-(MPC)

Maybe a pass through to check for these things is worthwhile.
