### 2022-10-10 #168 @moorepants
There isn't an issues tab on the new repository so I'll leave some things here. The old repository is still referenced in many places, for example:
![image](https://user-images.githubusercontent.com/276007/194848254-308fc7cf-0f63-44b2-b859-6a86c0438186.png)
Do you plan to update these?
