### 2022-10-10 #168 @moorepants
You are using a theme that has a dark mode. Many of the figures are unreadable, e.g.:
![image](https://user-images.githubusercontent.com/276007/194848584-f765223e-2426-4ce7-9637-5af18911769f.png)
Maybe disable the dark mode as a quick fix?
