### 2022-10-10 #168 @moorepants
Tutorial figure on line 7 has misplaced axes labels:
![image](https://user-images.githubusercontent.com/276007/194848793-99ed3ab0-406b-47fd-8150-d2bc57471720.png)
