### 2022-10-10 #168 @moorepants
The tutorial video on the last line does not fit on narrow screens:
![image](https://user-images.githubusercontent.com/276007/194848903-e51ad4cb-fc9c-460f-8446-a04292455590.png)
