### 2022-10-10 #168 @moorepants
The `sys.append('../')` lines should not be needed in your examples:
```python
import sys
sys.path.append('../')
%matplotlib notebook
from pendsim import sim, controller, viz
import numpy as np
import matplotlib.pyplot as plt
```
If users install your package, they shouldn't have to do that.
