### 2022-10-10 #168 @moorepants
On the animation page there is some unusual directory changing:
```
import os, pathlib
os.chdir(pathlib.Path(globals()['_dh'][0]).parent)
```
This does not seem like good practice.
