### 2022-10-10 #168 @moorepants
On the linearization page there are some formatting errors, e.g.:
![image](https://user-images.githubusercontent.com/276007/194849494-5c0ad716-1489-4e9b-8900-efd283090cc4.png)
