### 2022-10-10 #168 @moorepants
@EASEL-UCI Can you address the various formatting and presentation issues I've mentioned here? This would bring the contents to publishable quality with JOSE. Once that is done, you can cut a new release.

@repagh and @cooperrc, I assume that the "community guidelines" checkbox was not checked by mistake. Can you check that box if you are satisfied with the statement in the README?
