### 2022-10-11 43df719
Update install.rst
new repo link and directory name
