+++ setup.cfg
@@ -6,9 +6,9 @@ author_email = msutherl@uci.edu
 description = A pendulum simulator
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/rland93/pendsim
+url = https://github.com/easel-uci/pendsim
 project_urls =
-    Bug Tracker = https://github.com/rland93/pendsim/issues
+    Bug Tracker = https://github.com/easel-uci/pendsim/issues
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
