### 2023-01-19 e325e4e
Update sim.py
changed multiprocessing to multiprocess to work with IPython environments
