+++ requirements.txt
@@ -10,3 +10,4 @@ nbsphinx-link>=1.2
 nbsphinx>=0.8
 numpydoc>=1.3
 nb2plots>=0.6
+multiprocess<=0.70.14
