+++ docs/install.rst
@@ -1,23 +1,13 @@
 Install
 =======
 
-The package requires a >=Python 3.6 installation. I recommend using pip and the pre-packaged PyPi release, which will automatically resolve dependencies:
-
-.. code:: bash
-
-    pip install pendsim
-
-.. note::
-    It may be useful to create an empty virtual environment so that installation of this package does not interfere with system dependencies. For more information, see the `Python guide to virtual environments <https://docs.python.org/3/tutorial/venv.html>`_.
-
-
-Manual Installation
-~~~~~~~~~~~~~~~~~~~
-
-Alternatively, you can manually install pendsim directly through the git repository:
+The package requires a >=Python 3.6 installation. Manually install pendsim directly through the git repository:
 
 .. code:: bash 
 
     git clone https://github.com/easel-uci/pendsim.git
     cd pendsim
     pip install -r requirements.txt
+
+.. note::
+    It may be useful to create an empty virtual environment so that installation of this package does not interfere with system dependencies. For more information, see the `Python guide to virtual environments <https://docs.python.org/3/tutorial/venv.html>`_.
