+++ docs/conf.py
@@ -68,7 +68,7 @@ templates_path = ["_templates"]
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
-autodoc_mock_imports = ["scipy", "numpy", "matplotlib", "cvxpy", "filterpy"]
+autodoc_mock_imports = ["scipy", "numpy", "matplotlib", "cvxpy", "filterpy", "multiprocess"]
 numpydoc_show_class_members = False
 
 # -- Options for HTML output -------------------------------------------------
