### 2023-01-19 #168 @EASEL-UCI
@labarba Happy New Year. Thanks very much for the note.
@moorepants Thank you for noticing these remaining issues and for the detailed comments. We've made the following changes and believe we've addressed each of the comments:
- updated documentation so that we no longer reference the old repository
- fixed misplaced axis labels and improved figure labels throughout docs
- removed all sys, os, and path commands from examples
- fixed formatting of bulleted lists on linearization docs page
- executed all cells on PID page
- video scaling improved, and all videos can be expanded to fit a particular screen
- users can disable dark theme to better view figures in docs

The new release v1.2 is up-to-date with these changes.
