### 2023-02-01 #168 @moorepants
@EASEL-UCI this still shows documentation for 1.1: https://rland93.github.io/pendsim/index.html Should I be looking elsewhere?
