### 2023-02-03 #168 @moorepants
I've checked over everything again and I approve for publishing. @EASEL-UCI can you make the Zenodo archive of your version 1.2.0 and reply here with the DOI to that archive?
