### 2023-03-04 #168 @labarba
I notice that both reviewers left the "Community guidelines" item of the checklist unchecked. But they both approved for publication in a later comment, @cooperrc [here](https://github.com/openjournals/jose-reviews/issues/168#issuecomment-1259781500) and @repagh [here](https://github.com/openjournals/jose-reviews/issues/168#issuecomment-1270106675). I will proceed with publication, and mark those items as checked.
