### 2023-03-04 #168 @labarba
Congratulations, @rland93 and @EASEL-UCI, your JOSE paper is published! 🚀

Big thanks to our Editor: @moorepants, and to the Reviewers: @cooperrc, @repagh — your contributions make this adventure in publishing possible 🙏
